requirements.txt
flask
psutil
dnspython
requests
flask-login
werkzeug
reportlab
dnspython
python-whois